Add new endpoint for popularity endpoint request an id and int 1 / -1

+++ puzzle-showdown/src/main/java/com/codecool/puzzleshowdown/dto/puzzle/PuzzleDTO.java
@@ -1,4 +1,4 @@
 package com.codecool.puzzleshowdown.dto.puzzle;
 
-public record PuzzleDTO(String id, String table, String firstMove, float rating, float popularity) {
+public record PuzzleDTO(String id, String table, String firstMove, int rating, int popularity) {
 }
